DLS-11425: [GMP] Apply Scala 3 Update to gmp-frontend

+++ app/repositories/SingleCalculationSessionRepository.scala
@@ -45,7 +45,7 @@ class SingleCalculationSessionRepository @Inject()(
         Indexes.ascending("lastModified"),
         IndexOptions()
           .name("lastModifiedIdx")
-          .expireAfter(appConfig.cacheTtl, TimeUnit.SECONDS)
+          .expireAfter(appConfig.cacheTtl.toLong, TimeUnit.SECONDS)
       )
     )
   ) {
